pyproject.toml
[project]
name = "lulzcasz-dev"
version = "0.0.0"
description = "lulzcasz.dev source code"
readme = "README.md"
requires-python = "==3.14.3"
dependencies = [
    "beautifulsoup4==4.14.3",
    "celery==5.6.3",
    "django==5.2.12",
    "django-htmx==1.27.0",
    "django-storages[s3]==1.14.6",
    "django-tinymce==5.0.0",
    "gunicorn==25.3.0",
    "pillow==12.2.0",
    "psycopg==3.3.3",
    "redis==7.4.0",
    "whitenoise[brotli]==6.12.0",
]

[uv.tool]
package = false

[dependency-groups]
dev = [
    "ipython==9.12.0",
    "psycopg[binary]==3.3.3",
    "rust-just==1.48.1",
]
